refactor: extract fixtures to fixtures.py
