+++ workflow.py
@@ -1,23 +1,7 @@
-from pathlib import Path
 from types import SimpleNamespace
-from typing import Dict
 
-from fixtures import fixture
 from virtool_workflow import hooks, step
 from virtool_workflow.api.samples import SampleProvider
-
-
-@fixture
-def intermediate() -> SimpleNamespace:
-    return SimpleNamespace()
-
-
-@fixture
-def read_files(input_files: Dict[str, Path]):
-    return [
-        file.rename(f"reads_{n}.fq.gz")
-        for file, n in zip(input_files.values(), (1, 2))
-    ]
 
 
 @step(name="Run FastQC")
